Move GQL logic to gql_add_item_list #29
