+++ components/NewItemList.jsx
@@ -1,6 +1,5 @@
 import { useState } from 'react'
-import { API, graphqlOperation } from 'aws-amplify'
-import { createItemList } from '../src/graphql/mutations'
+import { gql_add_item_list } from '../src/graphql-actions'
 import {usePantry} from '../src/user-context'
 
 
@@ -13,27 +12,7 @@ export default function NewItemList () {
   const addList = async (event) => {
     event.preventDefault();
     try {
-      const graphqlResponse = await API.graphql(graphqlOperation(createItemList, {
-        input: {
-          title: event.target.title.value,
-          pantryDetails: {
-            ingredients: []
-          },
-          shoppingDetails: {
-            ingredients: []
-          }
-        }
-      }))
-      const newItemList = {
-        id: graphqlResponse.data.createItemList.id,
-        title: graphqlResponse.data.createItemList.title,
-        pantryDetails: {
-          ingredients: []
-        },
-        shoppingDetails: {
-          ingredients: []
-        }
-      }
+      const newItemList = await gql_add_item_list(event.target.title.value);
       dispatch({type: "ADD_ITEM_LIST", payload: newItemList})
       setTitle({
         title: "",
